fix(kernel): added full support for keyboard in user space

+++ kernel/include/kernel/proc/stream.h
@@ -13,8 +13,8 @@ typedef enum
     // STREAM_TYPE_BIDIRECTIONAL = 2
 } stream_type_t;
 
-typedef uint8_t (*stream_driver_read)(device_handle_t);
-typedef void (*stream_driver_write)(uint8_t, device_handle_t);
+typedef uint32_t (*stream_driver_read)(device_handle_t);
+typedef void (*stream_driver_write)(uint32_t, device_handle_t);
 
 typedef struct
 {
@@ -34,7 +34,7 @@ typedef struct
 stream_t *stream_create_driver(device_handle_t device, stream_driver_read read_func, stream_driver_write write_func);
 void stream_free(stream_t *stream);
 
-int stream_read(stream_t *stream, uint8_t *buf, size_t size);
-int stream_write(stream_t *stream, const uint8_t *buf, size_t size);
+int stream_read(stream_t *stream, uint32_t *buf, size_t size);
+int stream_write(stream_t *stream, const uint32_t *buf, size_t size);
 
 #endif
